% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{batchRename}
\alias{batchRename}
\title{Batch rename files}
\usage{
batchRename(folder = getwd(), filefilter = NULL, pattern, replacement)
}
\arguments{
\item{folder}{Path to the folder containing files to be renamed.}

\item{filefilter}{an optional regular expression. Only file names which match
the regular expression will be returned.}

\item{pattern}{regular expression describing the pattern in filenames to
be replaced.}

\item{replacement}{regular expression describing the pattern in filenames to
replace pattern with.}
}
\description{
Replace pattern with replacement for all files in the specified folder
matching the specified filefilter. Function is run for its side effect of
renaming files.
}
\examples{
file.create("1 a.txt")
file.create("1 b.txt")
batchRename(filefilter = "\\\\d \\\\w\\\\.txt", pattern = "\\\\s", replacement = "")
}
\author{
Caspar J. van Lissa
}
