% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{calc_i2}
\alias{calc_i2}
\title{Calculate I^2 for a multilevel meta-analysis}
\usage{
calc_i2(x)
}
\arguments{
\item{x}{An object of class \code{rma}.}
}
\value{
A numeric vector
}
\description{
Takes an object of class \code{rma}, containing a multi-level
meta-analysis with two variance components, and calculates the I^2 for
the variance components.
}
