% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_binaries}
\alias{combine_binaries}
\title{Generate a matrix of binary combinations}
\usage{
combine_binaries(n)
}
\arguments{
\item{n}{The number of combinations to produce.}
}
\value{
A matrix of binary combinations, one for each row.
}
\description{
Generates a matrix of all possible combinations of binary values (0s and 1s).
}
\examples{
combine_binaries(3)

}
\seealso{
\code{\link[base]{expand.grid}}, \code{\link[utils]{combn}}
}

