% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{compareLogisticModels}
\alias{compareLogisticModels}
\title{Compare logistic regression models}
\usage{
compareLogisticModels(models)
}
\arguments{
\item{models}{A (named) list of logistic regression models
(output from \code{lm}))}
}
\description{
Returns a model fit table with significance tests for a list of logistic
regression models.
}
\examples{
dat <- mtcars
model1 <- glm(am ~ mpg, data = dat, family =
"binomial")
model2 <- glm(am ~ mpg + wt, data = dat, family =
"binomial")
compareLogisticModels(models = list("Only mpg" = model1,
                                    "Mpg and wt" = model2))
}
\author{
Caspar J. van Lissa
}
