% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_functions.R
\name{count_references}
\alias{count_references}
\title{Count references (in text on clipboard)}
\usage{
count_references(x = readClipboard())
}
\arguments{
\item{x}{Character vector, Default: readClipboard()}
}
\value{
Object of class \code{table}
}
\description{
Counts the number of references (identified by the \code{@}
symbol) in a character vector, which is read from the clipboard by default.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 text <- c("Although sex differences in emotional adjustment are sometimes",
           "found in older children [@silkAdolescentsEmotionRegulation2003;",
           "@vanlissaRoleFathersMothers2018a], a meta-analysis found little",
           "evidence for such sex differences in young children",
           "[@else-questGenderDifferencesTemperament2006a].")
 count_references(text)
 }
}
}
