% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{doEFA}
\alias{doEFA}
\title{Conduct a full EFA analysis}
\usage{
doEFA(data, name, nfactors = NULL, write_files = FALSE)
}
\arguments{
\item{data}{Data.frame. Datafile to be factor analyzed.}

\item{name}{Character. The name of the output (for files to be written etc.)}

\item{nfactors}{Integer. Hypothesized number of factors; an EFA with this
number of factors will be included in the results.}

\item{write_files}{Logical. Whether to write results to .csv files or not.}
}
\description{
Conducts a full EFA analysis including assumption checks. Returns the
results as a list, writes to .csv files, and prints to the console.
}
\examples{
efa_results <- doEFA(bfi[1:500, 1:25], "bfi", 2)
}
\author{
Caspar J. van Lissa
}
