% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{escape_special}
\alias{escape_special}
\title{Escape special characters}
\usage{
escape_special(strings)
}
\arguments{
\item{strings}{Character vector.}
}
\value{
Character vector.
}
\description{
Takes in a string, and escapes any special characters so you can search for
the literal string using regex.
}
\examples{
escape_special("What the $^&(){}.*|?")
}
