% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{EtaSq}
\alias{EtaSq}
\title{Calculate partial eta squared for anova}
\usage{
EtaSq(x)
}
\arguments{
\item{x}{An anova model (from \code{aov()}).}
}
\description{
Takes in an anova model, and returns partial eta squared for the different
predictors.
}
\examples{
model1 <- aov(weight ~ group, data = PlantGrowth)
EtaSq(model1)
}
\author{
Caspar J. van Lissa
}
