% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{find_in_files}
\alias{find_in_files}
\title{Find text in files}
\usage{
find_in_files(text, max_size = 1e+05, dir = getwd(), ...)
}
\arguments{
\item{text}{Character vector. Literal text to search}

\item{max_size}{Numeric. Maximum file size to read.}

\item{dir}{Character. Which directory to search.}

\item{...}{Additional arguments, passed on to \code{list.files}.}
}
\value{
Nothing. Prints directly to the console.
}
\description{
Finds literal text in files. Useful if, for example, you want to replace a
function name in all files of an R-package, or if you want to see whether a
function can be deprecated because it is not used.
}
\examples{
# Make me!
}
\author{
Caspar J. van Lissa
}
