% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{funnel_plotlyfi}
\alias{funnel_plotlyfi}
\title{Create an interactive funnel plot}
\usage{
funnel_plotlyfi(x, labels = NULL)
}
\arguments{
\item{x}{An object of class \code{rma}.}

\item{labels}{A character vector with labels for the data points.}
}
\value{
A plotly plot.
}
\description{
Takes an object of class \code{rma}, and plots an interactive
funnel plot using \code{plotly}.
}
