% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{ifnofile}
\alias{ifnofile}
\title{Run if file does (not) exist, else load file}
\usage{
ifnofile(filename, expr)
}
\arguments{
\item{filename}{Check whether this file exists or not. If it does exist, load
the contents of that file into the global environment.}

\item{expr}{An expression to be evaluated if the filename does not exist.
Multiple expressions should be enclosed in curly braces. Objects created by
this expression are stored in a file called \code{filename}.}
}
\description{
Check if a file exists, and if it doesn't, run an expression, and save the
resulting objects to the specified filename. Useful when writing Rmarkdown
reports with time-intensive analyses; these can be run once and then saved.
}
\examples{
ifnofile("example.RData", {
  a <- 1
  b <- 2
})
ifnofile("example.RData", {
  a <- 1
  b <- 2
})
}
\author{
Caspar J. van Lissa
}
