% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{logisticPseudoR2s}
\alias{logisticPseudoR2s}
\title{Return pseudo R2s for a logistic model}
\usage{
logisticPseudoR2s(LogModel)
}
\arguments{
\item{logModel}{A logistic regression model (from \code{glm()}).}
}
\description{
Takes in a logistic regression model, and returns Hosmer & Lemeshow, Cox &
Snell, and Nagelkerke's pseudo-R2 estimates
}
\examples{
dat <- mtcars
model1 <- glm(am ~ mpg, data = dat, family =
"binomial")
logisticPseudoR2s(model1)
}
\author{
Andy Field
}
