% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{MplusConstrainModels}
\alias{MplusConstrainModels}
\title{Constrain Mplus syntax (old, redundant version)}
\usage{
MplusConstrainModels(mplusoutput = NULL, ztests = NULL,
  location = getwd(), outputlocation = getwd())
}
\arguments{
\item{mplusoutput}{modelList. A list of mplusModel outputs.}

\item{ztests}{A table of z-tests, returned by mplusToTable.}

\item{location}{Character. The location of relevant input files.}

\item{outputlocation}{Character. The location where to write the constrained
input files.}
}
\value{
Nothing. Function is run for its side effect of writing files.
}
\description{
This function was used for my paper on parenting and adolescents' emotion
regulation. It is intended to constrain non-significantly different
parameters across groups. Interesting idea, which could be implemented more
systematically using mplusModeler().
}
\examples{
#Make me!
}
\seealso{
\code{\link{readModels}}.

Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{conf_int}},
  \code{\link{corTable}}, \code{\link{est_sig}},
  \code{\link{mplusToTable}}, \code{\link{param_label}},
  \code{\link{printResultsTable}},
  \code{\link{rbind_tables}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
