% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{mplusToTable}
\alias{mplusToTable}
\title{Tabulate Mplus output (old, redundant version)}
\usage{
mplusToTable(location = getwd(), mplusoutput = NULL,
  results = "stdyx.standardized", se = FALSE)
}
\arguments{
\item{location}{Character. Location to read the Mplus output files from.}

\item{mplusoutput}{modelList. A list of mplusModel outputs.}

\item{results}{Character. Which section of the results to retain; one of
\code{c("unstandardized", "stdyx.standardized", "stdy.standardized")}.}

\item{se}{Logical. Report se or not.}
}
\value{
A list of Mplus results.
}
\description{
This function was used for my paper on parenting and adolescents' emotion
regulation. It is largely deprecated by printResultsTable, but contains some
interesting code for labeling Z-tests for new parameters, and tabulating
multi-group output. Should be re-examined and merged with printResultsTable.
}
\examples{
#Make me!
}
\seealso{
\code{\link{readModels}}.

Other Mplus functions: \code{\link{MplusConstrainModels}},
  \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{conf_int}},
  \code{\link{corTable}}, \code{\link{est_sig}},
  \code{\link{param_label}},
  \code{\link{printResultsTable}},
  \code{\link{rbind_tables}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
