% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize (scale) numeric vector or matrix to any range}
\usage{
normalize(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{A numeric vector or matrix}

\item{new_min}{The new min value for the input}

\item{new_max}{The new max value for the input}
}
\value{
A matrix or vector with values normalized to the given range.#'
}
\description{
Take a numeric vector or matrix and scale it to a new range.
}
\examples{
normalize(1:10, new_min = -10, new_max = -9)#'
}
\seealso{
\link[base]{scale}
}

