% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{poolSD}
\alias{poolSD}
\title{Pool standard deviations}
\usage{
poolSD(ns, sds)
}
\arguments{
\item{ns}{Vector of sample sizes for the standard deviations to pool.}

\item{sds}{Vector of standard deviations to pool.}
}
\value{
An atomic numeric vector.
}
\description{
Pool standard deviations
}
\examples{
poolSD(ns = c(64, 33), sds = c(.66, .78))
}
\author{
Caspar J. van Lissa
}
