% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{rb_from_rpb}
\alias{rb_from_rpb}
\title{Calculate biserial correlation from point-biserial correlation}
\usage{
rb_from_rpb(rpb, ne, nc)
}
\arguments{
\item{rpb}{F-value.}

\item{ne}{Sample size of the 'experimental condition'.}

\item{nc}{Sample size of the 'control condition'.}
}
\value{
An atomic numeric vector.
}
\description{
Calculate the biserial correlation from point-biserial correlation and sample
sizes. The biserial correlation is used when the dichotomous variable is an
artificial dichotomy. When the dichotomy is real, use the point-biserial
correlation (see \code{\link{rb_from_rpb}}).
}
\examples{
rb_from_rpb(rpb_from_f(2.44, 54, 65), 54, 65)
}
\seealso{
\code{\link{rpb_from_f}}
}
\author{
Caspar J. van Lissa
}
