% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{rpb_from_f}
\alias{rpb_from_f}
\title{Calculate point-biserial correlation from F-value and sample sizes}
\usage{
rpb_from_f(f, ne, nc)
}
\arguments{
\item{f}{F-value.}

\item{ne}{Sample size of the 'experimental condition'.}

\item{nc}{Sample size of the 'control condition'.}
}
\value{
An atomic numeric vector.
}
\description{
For a one-way ANOVA, calculate the point-biserial correlation from F-value
and sample sizes. The point-biserial correlation is used when the dichotomous
variable is a true dichotomy. When the dichotomy is artificial, use the
biserial correlation (see \code{\link{rb_from_rpb}}).
}
\examples{
rpb_from_f(2.44, 54, 65)
}
\seealso{
\code{\link{rb_from_rpb}}
}
\author{
Caspar J. van Lissa
}
