% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_chunk.R
\name{save_chunk}
\alias{save_chunk}
\title{Run chunk and save to file, load file if it exists}
\usage{
save_chunk(expr)
}
\arguments{
\item{expr}{An expression to be evaluated if the filename does not exist.
Multiple expressions should be enclosed in curly braces. Objects created by
this expression are stored in a file called \code{filename}.}
}
\description{
For an Rmarkdown chunk, check if a file exists containing the results,
and if it doesn't, run an expression, and save the resulting objects.
Useful when writing Rmarkdown
reports with time-intensive analyses; these can be run once and then saved.
}
\author{
Caspar J. van Lissa
}
