% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{screenResponses}
\alias{screenResponses}
\title{Screen data for nonsense answers}
\usage{
screenResponses(data = NULL, keys.list = NULL,
  maxpercentidentical = 1, maxquestionnaires = 2)
}
\arguments{
\item{data}{The dataset.}

\item{keys.list}{A (named) list, with one entry for each questionnaire,
containing a character vector with all items belonging to that questionnaire.}

\item{maxpercentidentical}{Numeric. Maximum allowed percentage of consecutive
identical answers within a questionnaire.}

\item{maxquestionnaires}{Maximum number of questionnaires allowed to exceed
maxpercentidentical.}
}
\value{
A list with elements \code{"screen"}: A logical vector of length
equal to the number of rows in \code{data}, identifying whether cases are
valid according to the thresholds specified; \code{"exceeded_by_person"}:
a numeric vector, indicating for how many questionnaires each person exceeded
\code{maxpercentidentical}; \code{"exceeded_by_questionnaire"}:
a numeric vector, indicating for each questionnaire how many persons exceeded
\code{maxpercentidentical}.
}
\description{
Experimental function, intended to flag nonsense responders in survey data.
}
\examples{
# Makeme!
}
\author{
Caspar J. van Lissa
}
