% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{summarySE}
\alias{summarySE}
\title{summarySE}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = TRUE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{measurevar}{Character. The name of the variable to summarize.}

\item{groupvars}{Character vector. Name of the grouping variables to
summarize by.}

\item{na.rm}{Logical. Remove NA values before summarizing?}

\item{conf.interval}{Numeric. What percentile to use for confidence interval?}

\item{.drop}{Logical. Drop unused levels of the grouping variables.}
}
\value{
A data.frame.
}
\description{
Takes in a data.frame, and returns summary statistics for the specified
\code{measurevar} by \code{groupvars}.
}
\examples{
dat <- iris
dat$Long <- cut(dat$Petal.Length, 2)
summarySE(dat, "Sepal.Length", c("Species", "Long"))
}
\author{
Caspar J. van Lissa
}
