% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{update_citekeys}
\alias{update_citekeys}
\title{Update citekeys using fuzzy string matching}
\usage{
update_citekeys(text, bibfile, ...)
}
\arguments{
\item{text}{Character string. Text from an Rmarkdown file.}

\item{bibfile}{Character string. Path to .bib file.}

\item{...}{Arguments passed to \code{\link[base]{agrep}}}
}
\value{
A character string.
}
\description{
When a new version of Zotero is released, sometimes, citekeys
are changed, thereby breaking references in an existing Rmarkdown file. This
function:
\enumerate{
\item Takes in a text string, and a file path to the .bib file
\item Extracts citekeys from both the text string and bib file
\item Uses fuzzy matching to replace the citekeys in text with the most
likely matching citekeys from bibfile.
}
}
\examples{
\dontrun{
update_citekeys(readLines("clipboard"), "references.bib")
}
}
