% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_tests.R
\name{z_test_coef}
\alias{z_test_coef}
\title{Conduct z-test for coefficients of a model}
\usage{
z_test_coef(x, par1, par2)
}
\arguments{
\item{x}{A model.}

\item{par1}{Numeric or character. Name or position of the first parameter.}

\item{par2}{Numeric or character. Name or position of the second parameter.}
}
\value{
Named vector.
}
\description{
Conduct z-test for coefficients of a model
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
