% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunctions.R
\name{screenResponses}
\alias{screenResponses}
\title{Screen data for nonsense answers}
\usage{
screenResponses(data, exclude = NULL, maxpercentidentical = 1,
  maxquestionnaires = 2, questionnaire_regex = "\\\\w+\\\\d+",
  questionnairelength = 3)
}
\arguments{
\item{data}{The dataset.}

\item{exclude}{Regex marker for questionnaires to exclude.}

\item{maxpercentidentical}{Maximum percentage of identical answers within
a questionnaire.}

\item{maxquestionnaires}{Maximum number of questionnaires which can exceed
maxpercentidentical.}

\item{questionnaire_regex}{Regex string that codes for a questionnaire.}

\item{questionnairelength}{Integer. How many questions should there be before
something is considered a questionnaire.}
}
\value{
A whole list of diagnostic stuff.
}
\description{
Experimental function, intended to flag nonsense responders in survey data.
}
\examples{
# Makeme!
}
\author{
Caspar J. van Lissa
}
